tweeter-web/src/components/authentication/login/Login.tsx
import "./Login.css";
import "bootstrap/dist/css/bootstrap.css";
import { useRef, useState } from "react";
import { Link, useNavigate } from "react-router-dom";
import AuthenticationFormLayout from "../AuthenticationFormLayout";
import { AuthToken, User } from "tweeter-shared";
import useToastListener from "../../toaster/ToastListenerHook";
import AutheticationField from "../authenticationFields/AuthenticationFields";
import useUserInfo from "../../userInfo/UserInfoHook";
import { LoginPresenter } from "../../../presenter/LoginPresenter";
import { AuthenticationView } from "../../../presenter/AuthenticationPresenter";

interface Props {
  originalUrl?: string;
  presenter?: LoginPresenter;
}

const Login = (props: Props) => {
  const [alias, setAlias] = useState("");
  const [password, setPassword] = useState("");
  const [rememberMe, setRememberMe] = useState(false);

  const navigate = useNavigate();
  const { updateUserInfo } = useUserInfo();
  const { displayErrorMessage } = useToastListener();

  const rememberMeRef = useRef(rememberMe);
  rememberMeRef.current = rememberMe;

  const listener:AuthenticationView = {
    displayErrorMessage: displayErrorMessage,
    navigate: navigate,
    updateUserInfo: (user: User, authToken: AuthToken) => updateUserInfo(user, user, authToken, rememberMeRef.current),
  }

  const checkSubmitButtonStatus = (): boolean => {
    return !alias || !password;
  };
  const [presenter] = useState(props.presenter ?? new LoginPresenter(listener))

  const doLogin = () => {
    presenter.doLogin(alias, password, props.originalUrl!)
  }

  const inputFieldGenerator = () => {
    return (
      <>
        <AutheticationField setAlias={(event) => setAlias(event.target.value)} setPassword= {(event) => setPassword(event.target.value)}/>
      </>
    );
  };

  const switchAuthenticationMethodGenerator = () => {
    return (
      <div className="mb-3">
        Not registered? <Link to="/register">Register</Link>
      </div>
    );
  };

  return (
    <AuthenticationFormLayout
      headingText="Please Sign In"
      submitButtonLabel="Sign in"
      oAuthHeading="Sign in with:"
      inputFieldGenerator={inputFieldGenerator}
      switchAuthenticationMethodGenerator={switchAuthenticationMethodGenerator}
      setRememberMe={setRememberMe}
      submitButtonDisabled={checkSubmitButtonStatus}
      submit={doLogin}
    />
  );
};

export default Login;
